+++ utils/marker/marker_factory.cpp
@@ -7,7 +7,7 @@ void MarkerFactory::InitMarker(Marker* m)
     ASSERT(m->GetGen() == Marker::GEN_UNSET);
 
     auto slots = Get()->slot_tracker_;
-    for (size_t i = 0; i < slots.size(); ++i) {
+    for (unsigned i = 0; i < slots.size(); ++i) {
         if (!slots[i]) {
             (Get()->slot_tracker_)[i] = true;
             auto gen = NewGen();
@@ -37,7 +37,7 @@ Marker::MarkerGenT MarkerFactory::NewGen()
     static auto current_gen = Marker::GEN_UNSET;
     // this will make so that GEN_UNSET will never be returned
     ++current_gen;
-    current_gen = current_gen + 1 * (current_gen == Marker::GEN_UNSET);
+    current_gen = current_gen + 1U * (current_gen == Marker::GEN_UNSET);
     return current_gen;
 }
