+++ utils/range/range.cpp
@@ -8,7 +8,7 @@ bool Range::IfIntersect(const Range& l, const Range& r)
     return !is_l_before_r && !is_r_before_l;
 }
 
-bool Range::Contains(size_t pt)
+bool Range::Contains(unsigned pt)
 {
     return start_ <= pt && pt < end_;
 }
@@ -23,10 +23,10 @@ Range Range::Intersection(const Range& l, const Range& r)
     ASSERT(IfIntersect(l, r));
 
     bool end_rel_pos = l.end_ < r.end_;
-    size_t end = end_rel_pos * l.end_ + !end_rel_pos * r.end_;
+    unsigned end = end_rel_pos * l.end_ + !end_rel_pos * r.end_;
 
     bool start_rel_pos = l.start_ > r.start_;
-    size_t start = start_rel_pos * l.start_ + !start_rel_pos * r.start_;
+    unsigned start = start_rel_pos * l.start_ + !start_rel_pos * r.start_;
 
     return Range(start, end);
 }
@@ -34,15 +34,15 @@ Range Range::Intersection(const Range& l, const Range& r)
 Range Range::Union(const Range& l, const Range& r)
 {
     bool end_rel_pos = l.end_ > r.end_;
-    size_t end = end_rel_pos * l.end_ + !end_rel_pos * r.end_;
+    unsigned end = end_rel_pos * l.end_ + !end_rel_pos * r.end_;
 
     bool start_rel_pos = l.start_ < r.start_;
-    size_t start = start_rel_pos * l.start_ + !start_rel_pos * r.start_;
+    unsigned start = start_rel_pos * l.start_ + !start_rel_pos * r.start_;
 
     return Range(start, end);
 }
 
-size_t Range::Length()
+unsigned Range::Length()
 {
     return end_ - start_;
 }
